Modified the package to work cross-platform
We have replaced the occurences of hard-coded file paths with the enviroment variable $UIO_APPLICATIONS.
The README.md describes how this should be set up prior to installation. We have also done the same with
the references to ROOT path -- this was hard-coded many places, but now everything is defined by the
enviroment variable $ROOTSYS.

NOTE: We have also commented out the make of 'rkinz', since we consider it to be superceded by the new program Qkinz (which may be found in a separate git repo on this account).

The README is also updated with a fix to a certain missing font problem which might occur on GNU/Linux platforms.

Fabio Zeiser and Jørgen E. Midtbø