Broader SiRi res to make sure to include wohle GS
Broader resolution for SiRi in normalization to make sure to include the whole GS. The broader resolution is important for eg. (He3,a) experiments and shouldn't have and adverse effect for eg. (d,p) experiements, where we have a better resolution. This is because in the latter case rho will be 0 anyhow in the region now included.
(Update from Magne)