Add possib. to include lowest 4 levels in gammex
Program to calculate <Gamma_g> as function of Ex. It is now possible to manually add the 4 lowest levels with spin and parity instead of using the NLD obtained with the Oslo method. Often the Oslo method fails to give correct NLD for the lowest levels, in particularly the 0+ ground state. In some cases this will give significant impact on the total <Gg>.
(Comitted for Magne)